Change to flex layout for ErrorView footer
